requirements.txt
fastapi[standard]==0.115.6
sqlmodel==0.0.22
pymysql==1.1.1
passlib==1.7.4
bcrypt==4.2.1
cryptography==44.0.0
PyJWT==2.10.1
fastapi-pagination==0.12.34